desktop/agentic_ai_udemy/Project-Autogen-main/requirements.txt
autogen
autogen-agentchat~=0.2
python-dotenv